Fix: [MacOS] manual sign libraries, as otherwise we miscalculate their checksums
Signing in CMake is done after the post-install hook, and there is
no hook-entry between signing creating the DMG. In result, we have
no way of calculating the checksums based on the signed libraries.

To resolve this, manually sign before we call cpack, after which
cpack signs the DMG (but not the libraries).